backend/requirements.txt
asgiref==3.10.0
cachetools==6.2.0
certifi==2025.10.5
charset-normalizer==3.4.3
dj-database-url==3.0.1
dj-rest-auth==6.0.0
Django==5.2.7
django-allauth==65.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gevent==25.9.1
google-api-core==2.26.0
google-auth==2.41.1
google-cloud-core==2.4.3
google-cloud-secret-manager==2.24.0
google-cloud-storage==3.4.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpc-google-iam-v1==0.14.2
grpcio==1.75.1
grpcio-status==1.75.1
gunicorn==23.0.0
idna==3.10
packaging==25.0
proto-plus==1.26.1
protobuf==6.32.1
psycopg==3.2.11
psycopg-binary==3.2.11
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.1.1
requests==2.32.5
rsa==4.9.1
setuptools==80.9.0
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
whitenoise==6.11.0
zope.event==6.0
zope.interface==8.0.1
